+++ web/src/pages/apd/ApdReadOnly.js
@@ -18,7 +18,7 @@ import PreviousActivities from './previous-activities/PreviousActivitiesReadOnly
 import Activities from '../../containers/viewOnly/activities/All';
 import ScheduleSummary from './schedule-summary/ScheduleSummaryReadOnly';
 import ProposedBudget from './proposed-budget/ProposedBudgetReadOnly';
-import AssuranceAndCompliance from '../../containers/viewOnly/AssurancesAndCompliance';
+import AssuranceAndCompliance from './assurances-and-compliance/AssurancesAndComplianceReadOnly';
 import ExecutiveSummary from './executive-summary/ExecutiveSummaryReadOnly';
 import ExportInstructions from '../../containers/viewOnly/Export';
 import Loading from '../../components/Loading';
